Make pip install editables to /opt
Otherwise it will try to put them in a subdirectory (src/) of
the source tree. This seems to work on terraform but the docker-compose
build seems to overwrite it so the bootstrap script can't run.

This fix also means we don't end up with a src/ directory added to the
source tree we're working with.