Rename S3 fields
This preserves the S3 field names in the Wellcome fork, which are now
different from the ones in the core repository.

This is necessary after rebasing the Wellcome fork against the latest
core (pre-1.11)

e.g. we originally renamed `access_key_id` to `aws_access_key_id` (in
an attempt to share the s3 fields and functionality between s3 and
wellcome plugins). In the S3 PR that was merged into core, it is still
named `access_key_id`. But we have live databases with columns named
`aws_access_key_id`. So for now, in this fork, we rename the fields back
again so as not to require fiddly column renaming. (this applies to
multiple fields).

We may want to rename them back again later on to bring everything
closer to core, but that's better done when not also trying to rebase
and also hand over the code :)